rename token switch file
